タイムライン画面に友達のカプセル表示画面
